Collapsed dotValue() into getting()
In the background to these commits I've been adding docblocks with full
descriptions for every method. I found it very difficult to describe in
plain english, it had very little meaning and didn't do a great deal.
All it did was extract an element from an array and that appropriate
exceptions would be thrown on key overflow or simply not existing.

I also realised that I was only calling it from one method, getting(),
though it was being called from two positions within that one method.

It really only existed because otherwise I would have repeated code.
That's why I couldn't describe it, it was just a container for some
repeated lines.

Well it turns out that it doesn't actually need to be called from two
positions in the method at all!